feat: 增加标签查询

+++ src/main/sql/3.classify.sql
@@ -1,6 +1,6 @@
 -- 专题表
 CREATE TABLE `blog_classify`(
-   `id` bigint(10) AUTO_INCREMENT,
+   `id` int(10) AUTO_INCREMENT,
    `gmt_create` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
    `gmt_modify` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '修改时间',
    `desc` varchar(240) NOT NULL COMMENT '专题简介',
