+++ src/main/sql/6.books.sql
@@ -1,6 +1,6 @@
 -- 个人书单表
 CREATE TABLE `blog_books` (
-  `id` bigint(11) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键',
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键',
   `progress` int(1) NOT NULL DEFAULT 0 COMMENT '状态: 0：未读, 1: 正在读, 2：已读',
   `classify_id` int(2) NOT NULL DELETE 0 COMMENT '数的分类标签',
   `book_name` varchar(100) NOT NULL  COMMENT '书名',
