+++ src/main/sql/7.maxim.sql
@@ -1,6 +1,6 @@
 -- 个人箴言表
 CREATE TABLE `blog_maxim` (
-  `id` bigint(11) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键',
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键',
   `state` int(1) NOT NULL DEFAULT 1 COMMENT '是否显示: 0：隐藏，1：显示',
   `context` varchar(520) NOT NULL  COMMENT '箴言内容',
   `gmt_create` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
